+++ src/Products/PloneMeeting/browser/advices.py
@@ -31,7 +31,6 @@ from plone.memoize import ram
 from Products.CMFCore.permissions import ModifyPortalContent
 from Products.CMFCore.utils import _checkPermission
 from Products.Five import BrowserView
-from Products.PloneMeeting.interfaces import IMeetingItem
 from zope.event import notify
 from zope.lifecycleevent import ObjectModifiedEvent
 
@@ -132,16 +131,6 @@ class AdvicesIcons(BrowserView):
             if advice_portal_type not in res:
                 res.append(advice_portal_type)
         return res
-
-    def get_tooltipster_base_class(self):
-        """Return a different CSS class if context is a dashboard or not
-           so we may initialize the tooltipster differently (regarding 'position' especially)."""
-        published = self.request.get('PUBLISHED', None)
-        published = published and published.aq_parent or self.context
-        base_class = 'tooltipster-dashboard-advices-infos'
-        if IMeetingItem.providedBy(published):
-            base_class = 'tooltipster-advices-infos'
-        return base_class
 
 
 class AdvicesIconsInfos(BrowserView):