+++ src/Products/PloneMeeting/profiles/__init__.py
@@ -448,7 +448,8 @@ class MeetingConfigDescriptor(Descriptor):
                          'hideNotViewableLinkedItemsTo', 'inheritedAdviceRemoveableByAdviser', 'usingGroups',
                          'hideHistoryTo')
     excludedFields = ['maxDaysDecisions', 'meetingAppDefaultView',
-                      'addContactsCSV', 'orderedContacts', 'orderedItemInitiators', 'disabled_collections']
+                      'addContactsCSV', 'orderedContacts', 'orderedItemInitiators',
+                      'disabled_collections', 'defaultLabels']
 
     # The 'instance' static attribute stores an instance used for assigning
     # default values to a meeting config being created through-the-web.
@@ -532,6 +533,10 @@ class MeetingConfigDescriptor(Descriptor):
             "here.restrictedTraverse('pm_unrestricted_methods').getLinkedMeetingDate().strftime('%Y%m%d') or '') " \
             "+ '/' + str(here.getItemNumber(relativeTo='meeting', for_display=True))"
         self.enableLabels = False
+        # labels are like :
+        # {'read': {'color': 'blue', 'label_id': 'read', 'by_user': True, 'title': 'Read'},
+        #  'urgent': {'color': 'red', 'label_id': 'urgent', 'by_user': False, 'title': 'Urgent'}}}
+        self.defaultLabels = {}
         # When adding items to a meeting, what sortingMethod must be applied successively?
         self.insertingMethodsOnAddItem = ({'insertingMethod': 'at_the_end', }, )
         # List if item tags defined for this meeting config